// frontend/src/lib/api/models/BenchmarkRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BenchmarkRead = {
    metric: string;
    studentAverage: (string | null);
    classAverage: (string | null);
    cohortSize: number;
    suppressed: boolean;
    suppressionMinCohort: number;
};