// frontend/src/lib/api/models/ForecastRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ForecastRead = {
    state: string;
    label: string;
    targetLetterGrade: string;
    targetPoints: string;
    earnedSoFar: string;
    remainingWeight: string;
    minReachable: string;
    maxReachable: string;
    currentLetterGrade: string;
    bestReachableLetterGrade: string;
    requiredRemainingAverage: (string | null);
    finalLetterGrade?: (string | null);
};