// frontend/src/lib/api/models/ProgressModuleSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProgressModuleSummary = {
    moduleId: string;
    title: string;
    currentStanding: (string | null);
    currentLetterGrade: (string | null);
    targetLetterGrade: (string | null);
    forecastState: (string | null);
    forecastLabel: (string | null);
    latestWeekNumber: (number | null);
    latestStandingPoints: (string | null);
};